// src/app/pending-verification/model/user.ts
export class User {
  id; string;
  isActive = false;
  name: string;
  last_name: string;
  email: string;
  ip: string;
  score: number;
  motivation: string;
  feedback: string;
  uses_vpn: boolean;
  is_first_account: boolean;
  is_first_application: boolean;
  fingerprint: string;
  ticket_id: string;
  browser: string;
  os: string;
  device: string;
  status: number;
  created_at: Date;
  updated_at: Date;

}