// src/app/pending-verification/model/twitter.ts
export class Twitter {
  user_id: number;
  name: string;
  screen_name: string;
  email: string;
  twitter_created_at: Date;
  followers_count: number;
  friends_count: number;
  statuses_count: number;
  location: string;
  profile_image_url: string;
  description: string;
  created_at: Date;
  updated_at: Date;
}